feat: move purchase domestic UI to components
